% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_fn.R
\name{control_fn}
\alias{control_fn}
\title{Create list of control options (to pass to \code{emuFit()})}
\usage{
control_fn(
  control = list(),
  max_step = 1,
  ignore_stop = FALSE,
  use_fullmodel_info = FALSE,
  use_fullmodel_cov = FALSE,
  use_both_cov = FALSE,
  inner_maxit = 25,
  inner_tol = 1,
  c1 = 1e-04,
  trackB = FALSE,
  return_nullB = FALSE,
  return_score_components = FALSE,
  return_both_score_pvals = FALSE,
  B_null_tol = 0.001,
  rho_init = 1,
  tau = 2,
  kappa = 0.8,
  constraint_tol = 1e-05,
  ntries = 4
)
}
\arguments{
\item{control}{Current control list (optional), will augment it with missing arguments}

\item{max_step}{Maximum stepsize; update directions computed during estimation (under the alternative).
Will be rescaled if a step in any parameter exceeds this value. Defaults to 1.}

\item{ignore_stop}{whether to ignore stopping criteria and run \code{maxit} iterations (helpful for diagnostic plots to determine convergence).}

\item{use_fullmodel_info}{Used in estimation under the null hypothesis. Whether to use information matrix from
estimation under the alternative hypothesis to construct the robust score statistic (instead of information
matrix from estimation under the null hypothesis). Defaults to \code{FALSE}.}

\item{use_fullmodel_cov}{Used in estimation under the null hypothesis. Whether to use covariance matrix from
estimation under the alternative hypothesis to construct the robust score statistic (instead of covariance
matrix from estimation under the null hypothesis). Defaults to \code{FALSE}.}

\item{use_both_cov}{Used in estimation under the null hypothesis. Whether to do score test twice, once with
covariance matrix under the alternative hypothesis and once with covariance matrix under the null hypothesis.
Defaults to \code{FALSE}.}

\item{inner_maxit}{Used in estimation under the null hypothesis. Maximum number of iterations within each inner loop
of estimation under null hypothesis algorithm. Default is \code{25}.}

\item{inner_tol}{Used in estimation under the null hypothesis. Convergence tolerance within each inner loop
of estimation under null hypothesis algorithm. Default is \code{1}.}

\item{c1}{Used in estimation under the null hypothesis. Parameter for Armijo line search. Default is \code{1e-4}.}

\item{trackB}{Used in estimation under the null hypothesis. When \code{TRUE} will track the value of \code{B} in each iteration
of optimization algorithm. Defaults to \code{FALSE}.}

\item{return_nullB}{Used in estimation under the null hypothesis. When \code{TRUE} will return the final value of \code{B}
under each null hypothesis tested. Defaults to \code{FALSE}.}

\item{return_score_components}{Used in estimation under the null hypothesis. When \code{TRUE} will return the components
of the robust score test statistic for each null hypothesis tested. Defaults to \code{FALSE}.}

\item{return_both_score_pvals}{Used in estimation under the null hypothesis, with \code{use_both_cov}. Defaults to \code{FALSE}.}

\item{B_null_tol}{Used in estimation under the null hypothesis, for the augmented Lagrangian algorithm.
numeric: convergence tolerance for null model fits for score testing (if max of absolute difference in
B across outer iterations is below this threshold, we declare convergence). Default is \code{0.001}.}

\item{rho_init}{Used in estimation under the null hypothesis, for the augmented Lagrangian algorithm.
Value at which to initiate rho parameter in augmented Lagrangian algorithm. Default is \code{1}.}

\item{tau}{Used in estimation under the null hypothesis, for the augmented Lagrangian algorithm.
Value to scale \code{rho} by in each iteration of augmented Lagrangian algorithm that does not move estimate
toward zero sufficiently. Default is \code{2}.}

\item{kappa}{Used in estimation under the null hypothesis, for the augmented Lagrangian algorithm. Value
between \code{0} and \code{1} that determines the cutoff on the ratio of current distance from feasibility over
distance in last iteration triggering scaling of \code{rho}. If this ratio is above \code{kappa}, \code{rho} is scaled
by \code{tau} to encourage estimate to move toward feasibility.}

\item{constraint_tol}{Used in estimation under the null hypothesis, for the augmented Lagrangian algorithm.
Constraint tolerance for fits under null hypotheses (tested element of \code{B} must be equal to constraint
function to within this tolerance for a fit to be accepted as a solution to constrained optimization
problem). Default is \code{1e-5}.}

\item{ntries}{Used in estimation under the null hypothesis, for the augmented Lagrangian algorithm.
The number of times to try optimization. Successive tries will change \code{tau} and \code{inner_maxit} and retry.}
}
\value{
A list containing control options, to have more control over optimization algorithms used by \code{radEmu}.
This can be passed into \code{emuFit()}.
}
\description{
Create list of control options (to pass to \code{emuFit()})
}
