% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpseudohuber_median_dx.R
\name{dpseudohuber_median_dx}
\alias{dpseudohuber_median_dx}
\alias{dpsuedohuber_median_dx}
\title{Calculate the derivative to the pseudo-Huber smoothed median}
\usage{
dpseudohuber_median_dx(x, d = 0.1)

dpsuedohuber_median_dx(x, d = 0.1)
}
\arguments{
\item{x}{A vector to calculate the derivative of the pseudo-Huber smoothed median for.}

\item{d}{Smoothing parameter, by default set to \code{0.1}. As \code{d} approaches
\code{0} the pseudo-Huber median function approaches the median and as \code{d} approaches
infinity this function approaches the mean.}
}
\value{
The derivative of the calculated pseudo-Huber smoothed median over \code{x} with smoothing
parameter \code{d}.
}
\description{
Calculate the derivative to the pseudo-Huber smoothed median
}
\examples{
dpseudohuber_median_dx(x = rnorm(10), d = 0.1)

}
