% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuCI.R
\name{emuCI}
\alias{emuCI}
\title{Construct confidence intervals for parameters estimated in fitted model}
\usage{
emuCI(
  emuMod,
  nboot,
  conf_level = 0.95,
  rows_of_interest = NULL,
  parallel = FALSE,
  blocks = NULL,
  verbose = FALSE,
  type = "bootstrap_percentile",
  nsim = 1000,
  ncore = 1
)
}
\arguments{
\item{emuMod}{An emuMod object fit with \code{emuFit}.}

\item{type}{The type of standard error estimation approach. Options are
\code{bayesian_subsample}, \code{sandwich}, and \code{bootstrap_percentile}.}
}
\value{
A confidence interval for each \code{beta} parameter estimated
in the \code{emuMod}.
}
\description{
This function takes in a radEmu model object and constructs confidence
intervals. Intervals can be constructed using a variety of robust
standard error estimation methods.
}
\author{
David Clausen
}
