% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuFit_check.R
\name{emuFit_check}
\alias{emuFit_check}
\title{Runs checks for appropriate arguments before running \code{emuFit()}}
\usage{
emuFit_check(
  Y,
  X = NULL,
  formula = NULL,
  data = NULL,
  assay_name = NULL,
  cluster = NULL,
  B_null_list = NULL,
  test_kj = NULL,
  match_row_names = TRUE,
  verbose = FALSE,
  remove_zero_comparison_pvals = 0.01,
  unobserved_taxon_error = TRUE,
  constraint_fn,
  constraint_grad_fn,
  constraint_param,
  run_score_tests = TRUE,
  null_fit_alg = "constraint_sandwich"
)
}
\arguments{
\item{Y}{an n x J matrix or dataframe of nonnegative observations, or a phyloseq object containing an otu table and sample data.}

\item{X}{an n x p matrix or dataframe of covariates (optional)}

\item{formula}{a one-sided formula specifying the form of the mean model to be fit}

\item{data}{an n x p data frame containing variables given in \code{formula}}

\item{assay_name}{a string containing the desired assay name within a \code{TreeSummarizedExperiment} object.
This is only required if Y is a \code{TreeSummarizedExperiment} object, otherwise this argument does nothing
and can be ignored.}

\item{cluster}{a vector giving cluster membership for each row of Y to be used in computing
GEE test statistics. Default is NULL, in which case rows of Y are treated as independent.}

\item{B_null_list}{list of starting values of coefficient matrix (p x J) for null estimation. This should either
be a list with the same length as \code{test_kj}. If you only want to provide starting values for some tests,
include the other elements of the list as \code{NULL}.}

\item{test_kj}{a data frame whose rows give coordinates (in category j and
covariate k) of elements of B to construct hypothesis tests for. \code{k} could also be the name of a covariate
included in \code{X} or \code{data}. If you don't know which coordinates k correspond to the covariate(s) that you
would like to test, run the function \code{radEmu::make_design_matrix()} in order to view the design matrix,
and identify which column of the design matrix corresponds to each covariate in your model. This argument
is required when running score tests.}

\item{match_row_names}{logical: Make sure rows on covariate data and response data correspond to
the same sample by comparing row names and subsetting/reordering if necessary.}

\item{verbose}{provide updates as model is being fitted? Defaults to FALSE. If user sets verbose = TRUE,
then key messages about algorithm progress will be displayed. If user sets verbose = "development",
then key messages and technical messages about convergence will be displayed. Most users who want status
updates should set verbose = TRUE.}

\item{remove_zero_comparison_pvals}{Should score p-values be replaced with NA for zero-comparison parameters? These parameters occur
for categorical covariates with three or more levels, and represent parameters that compare a covariate level to the reference level for
a category in which the comparison level and reference level both have 0 counts in all samples. These parameters can have misleadingly
small p-values and are not thought to have scientifically interesting signals. We recommend removing them before analyzing data further.
If TRUE, all zero-comparison parameter p-values will be set to NA. If FALSE no zero-comparison parameter p-values will be set to NA.
If a value between 0 and 1, all zero-comparison p-values below the value will be set to NA.
Default is \code{0.01}.}

\item{unobserved_taxon_error}{logical: should an error be thrown if Y includes taxa that have 0 counts for all samples? Default is TRUE.}

\item{constraint_fn}{function g defining a constraint on rows of B; g(B_k) = 0
for rows k = 1, ..., p of B. Default function is a smoothed median (minimizer of
pseudohuber loss). If a number is provided a single category constraint will be used
with the provided category as a reference category. This argument can either be a single constraint
function to be used for all rows of B, or a list of length p of constraints to be used for each row of B.}

\item{constraint_grad_fn}{derivative of constraint_fn with respect to its
arguments (i.e., elements of a row of B). If \code{constraint_fn} is a list of constraint functions, then
this argument must also be a list.}

\item{constraint_param}{If pseudohuber centering is used (this is the default),
parameter controlling relative weighting of elements closer and further from center.
(Limit as \code{constraint_param} approaches infinity is the mean; as this parameter approaches zero,
the minimizer of the pseudo-Huber loss approaches the median.)}

\item{run_score_tests}{logical: perform robust score testing?}

\item{null_fit_alg}{Which null fitting algorithm to use \code{"fisher_scoring"} or \code{"augmented_lagrangian"}.
Default and recommended approach is \code{"fisher_scoring"}.}
}
\value{
returns objects \code{Y}, \code{X}, \code{cluster}, and \code{B_null_list}, which may be modified by tests, and throw any useful
errors, warnings, or messages.
}
\description{
Runs checks for appropriate arguments before running \code{emuFit()}
}
