% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null.R
\name{fit_null}
\alias{fit_null}
\title{Fit model under null}
\usage{
fit_null(
  B,
  Y,
  X,
  X_cup = NULL,
  k_constr,
  j_constr,
  j_ref,
  constraint_fn,
  constraint_grad_fn,
  rho_init = 1,
  tau = 1.2,
  kappa = 0.8,
  B_tol = 0.01,
  inner_tol = 0.01,
  constraint_tol = 1e-04,
  max_step = 5,
  c1 = 1e-04,
  maxit = 1000,
  inner_maxit = 25,
  n_final_inner_it = 25,
  verbose = FALSE,
  trackB = FALSE
)
}
\description{
Fit model under null
}
