% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null_discrete.R
\name{fit_null_discrete}
\alias{fit_null_discrete}
\title{fits model with B_kj constrained to equal g(B_k) for constraint fn g, for a symmetric constraint with a discrete design}
\usage{
fit_null_discrete(
  Y,
  X,
  k_constr,
  j_constr,
  j_ref,
  trackB = FALSE,
  maxit = 5000,
  tol = 0.1,
  verbose = FALSE,
  ls_max = 20,
  ls_rho = 0.5,
  max_step = 1,
  constraint = "pseudohuber"
)
}
\arguments{
\item{Y}{Y (with augmentations)}

\item{X}{design matrix}

\item{k_constr}{row index of B to constrain}

\item{j_constr}{col index of B to constrain}

\item{j_ref}{column index of convenience constraint}

\item{trackB}{track value of beta across iterations and return? Default is \code{FALSE}.}

\item{maxit}{maximum iterations. Default is \code{5000}.}

\item{tol}{tolerance for stopping criteria. Algorithm stops when the root mean of the
norm of the score vector is less than the tolerance. Default is \code{0.01}.}

\item{verbose}{should the algorithm print updates for you? Default is \code{FALSE}.}

\item{ls_max}{maximum number of iterations in  the line search. Default is \code{20}.}

\item{ls_rho}{scaling factor in the line search. Default is \code{0.5}.}

\item{max_step}{step capping after the line sesarch. Default is \code{1}.}

\item{constraint}{What type of symmetric constraint do we have? Options are \code{"mean"} and \verb{"pseudohuber}.}
}
\value{
A list containing elements \code{B}, \code{k_constr}, \code{j_constr}, \code{niter}
and \code{Bs}. \code{B} is a matrix containing parameter estimates
under the null (obtained by maximum likelihood on augmented observations Y),
\code{k_constr}, and \code{j_constr} give row and column indexes of the parameter
fixed to be equal to the constraint function \eqn{g()} under the null. \code{niter} is a
scalar giving total number of outer iterations used to fit the null model, and
\code{Bs} is a data frame containing values of B by iteration if \code{trackB} was set
equal to TRUE (otherwise it contains a NULL value).
}
\description{
fits model with B_kj constrained to equal g(B_k) for constraint fn g, for a symmetric constraint with a discrete design
}
