% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null_symmetric_subset.R
\name{fit_null_symmetric_subset}
\alias{fit_null_symmetric_subset}
\title{fits model with B_kj constrained to equal g(B_k) for constraint fn g, for a symmetric constraint}
\usage{
fit_null_symmetric_subset(
  B,
  Y,
  X,
  X_cup = NULL,
  k_constr,
  j_constr,
  j_ref,
  constraint_fn,
  constraint_grad_fn,
  B_tol = 0.01,
  inner_tol = 0.01,
  c1 = 0.01,
  maxit = 1000,
  inner_maxit = 25,
  verbose = FALSE,
  trackB = FALSE,
  use_optim = FALSE,
  ignore_stop = FALSE,
  tol_lik = 1e-05,
  tol_test_stat = 0.01,
  null_window = 5,
  max_step = 1,
  reference_set = NULL
)
}
\arguments{
\item{B}{description}

\item{Y}{Y (with augmentations)}

\item{X}{design matrix}

\item{X_cup}{design matrix for Y in long format. Defaults to NULL, in which case matrix is computed from X.}

\item{k_constr}{row index of B to constrain}

\item{j_constr}{col index of B to constrain}

\item{j_ref}{column index of convenience constraint}

\item{constraint_fn}{constraint function}

\item{constraint_grad_fn}{gradient of constraint fn}

\item{B_tol}{tolerance for convergence in \eqn{max_{k,j} \lvert B^t_{kj} - B^{(t - 1)}_{kj}\rvert}}

\item{inner_tol}{tolerance for inner loop}

\item{c1}{constant for armijo rule}

\item{maxit}{maximum iterations}

\item{inner_maxit}{max iterations per inner loop}

\item{verbose}{shout at you?}

\item{trackB}{track value of beta across iterations and return?}

\item{use_optim}{whether to use \code{optim} instead of fisher scoring. Default is FALSE.}

\item{ignore_stop}{whether to ignore stopping criteria and run \code{maxit} iterations (could be helpful for diagnostic plots).}

\item{tol_lik}{tolerance for relative changes in likelihood for stopping criteria. Default is \code{1e-5}.}

\item{tol_test_stat}{tolerance for relative changes in test statistic for stopping criteria. Default is \code{0.01}.}

\item{null_window}{window to use for stopping criteria (this many iterations where stopping criteria is met). Default is \code{5}.}

\item{max_step}{Default is \code{1}.}

\item{reference_set}{Optional reference set to act as the subset. Default is \code{NULL} in which case the function will look into
\code{constraint_fn} for the reference set.}
}
\value{
A list containing elements \code{B}, \code{k_constr}, \code{j_constr}, \code{niter}
and \code{Bs}. \code{B} is a matrix containing parameter estimates
under the null (obtained by maximum likelihood on augmented observations Y),
\code{k_constr}, and \code{j_constr} give row and column indexes of the parameter
fixed to be equal to the constraint function \eqn{g()} under the null. \code{niter} is a
scalar giving total number of outer iterations used to fit the null model, and
\code{Bs} is a data frame containing values of B by iteration if \code{trackB} was set
equal to TRUE (otherwise it contains a NULL value). - update based on new algorithm
}
\description{
fits model with B_kj constrained to equal g(B_k) for constraint fn g, for a symmetric constraint
}
