% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_matrix.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Generates the design matrix for a given \code{formula} and \code{data} object}
\usage{
make_design_matrix(Y = NULL, formula, data = NULL, assay_name = NULL)
}
\arguments{
\item{Y}{Optionally, a \code{phyloseq} or \code{TreeSummarizedExperiment} object
containing an otu table and sample data.}

\item{formula}{a one-sided formula specifying the form of the mean model to be fit}

\item{data}{an n x p data frame containing variables given in \code{formula} (required unless
\code{Y} is included as a \code{phyloseq} or \code{TreeSummarizedExperiment} object)}

\item{assay_name}{Optionally, a string containing the desired assay name within a \code{TreeSummarizedExperiment} object.
This is only required if \code{data} is a \code{TreeSummarizedExperiment} object, otherwise this argument does nothing
and can be ignored.}
}
\value{
returns design matrix \code{X}.
}
\description{
Generates the design matrix for a given \code{formula} and \code{data} object
}
