% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagnostics_plots.R
\name{make_diagnostics_plots}
\alias{make_diagnostics_plots}
\title{makes plots to investigate convergence of estimation under null hypothesis}
\usage{
make_diagnostics_plots(diagnostic_df)
}
\arguments{
\item{diagnostic_df}{Dataframe with relevant information}
}
\value{
Plots with diagnostic information. If the fisher scoring algorithm has been fit, this will be two
plots showing how the log likelihood and test statistic change over iterations. If the augmented lagrangian
algorithm has been fit, this will also include how the constraint gap and maximum changing element of B change
over iterations.
}
\description{
makes plots to investigate convergence of estimation under null hypothesis
}
