% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reference_constraints.R
\name{make_reference_constraints}
\alias{make_reference_constraints}
\title{Make lists of constraint functions and their gradients when using a reference taxon}
\usage{
make_reference_constraints(p, j)
}
\arguments{
\item{p}{The number of columns in the design matrix \code{X}. If you don't know the
number of columns, you can find it with
\code{ncol(radEmu::make_design_matrix(formula = your_formula, data = your_data))}.
\code{your_formula} should be the expression you give to \code{emuFit}'s
\code{formula} argument.}

\item{j}{A single value or a vector of length \code{p - 1} where \code{p} is the number
of columns in the design matrix \code{X}. If a single value, \code{j} will be used as
the reference category for all covariates. If a vector of values, \code{j[k]} will be
used as the reference category for the covariate in design matrix column \code{k + 1}.}
}
\value{
A list with elements \code{constraints_list} and \code{constraints_grad_list}.
The \code{constraints_list} is a list of constraint functions for each column \code{p}
of the design matrix. By default, the constraint for the intercept is the pseudo Huber
median. The constraints for covariates are determined by reference categories given by
the argument \code{j}. The \code{constraints_grad_list} is a list of gradients of each
constraint function.
}
\description{
Make lists of constraint functions and their gradients when using a reference taxon
}
\examples{
# two columns in design matrix, reference taxon is taxon 5
list1 <- make_reference_constraints(p = 2, j = 5)

# four columns in design matrix, reference taxon for all covariates is taxon 2
list2 <- make_reference_constraints(p = 4, j = 2)

# four columns in design matrix, reference taxon for covariates 1 and 2 is taxon 3 and
# reference taxon for covariate 3 is taxon 4
list3 <- make_reference_constraints(p = 4, j = c(3, 3, 4))

}
