% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emuFit.R
\name{print.emuFit}
\alias{print.emuFit}
\title{Print function}
\usage{
\method{print}{emuFit}(x, n = 20, ...)
}
\arguments{
\item{x}{Object of class \code{emuFit}}

\item{n}{The number of coefficient estimates to be printed (ordered by largest absolute value to smallest)}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
\code{NULL}. Displays printed model summary.
}
\description{
Print function
}
