% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudohuber_median.R
\name{pseudohuber_median}
\alias{pseudohuber_median}
\alias{psuedohuber_median}
\title{Calculate the pseudo-Huber smoothed median}
\usage{
pseudohuber_median(x, d = 0.1, tolerance = 1e-08)

psuedohuber_median(x, d = 0.1, tolerance = 1e-08)
}
\arguments{
\item{x}{A vector to calculate the pseudo-Huber smoothed median for.}

\item{d}{Smoothing parameter, by default set to \code{0.1}. As \code{d} approaches
\code{0} this function approaches the median and as \code{d} approaches infinity
this function approaches the mean.}

\item{tolerance}{Tolerance used to determine convergence in the algorithm used
to calculate this function value.}
}
\value{
The calculated pseudo-Huber smoothed median over \code{x} with smoothing
parameter \code{d}.
}
\description{
Calculate the pseudo-Huber smoothed median using a quadratic approximation to the
pseudo-Huber criterion detailed in supplement to Clausen et al. (2024).
}
\examples{
pseudohuber_median(x = rnorm(10), d = 0.1)

}
