% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radEmu-package.R
\docType{package}
\name{radEmu}
\alias{radEmu-package}
\alias{radEmu}
\title{radEmu: Using Relative Abundance Data to Estimate of Multiplicative Differences in Mean Absolute Abundance}
\description{
The radEmu package provides a method for estimating and testing multiplicative differences in mean absolute abundance, using relative abundance data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statdivlab/radEmu}
  \item \url{https://statdivlab.github.io/radEmu/}
}

}
\author{
\strong{Maintainer}: Amy D Willis \email{adwillis@uw.edu} (\href{https://orcid.org/0000-0002-2802-4317}{ORCID})

Authors:
\itemize{
  \item Amy D Willis \email{adwillis@uw.edu} (\href{https://orcid.org/0000-0002-2802-4317}{ORCID})
  \item David Clausen
  \item Sarah Teichman
}

}
\keyword{internal}
