% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Data simulation function}
\usage{
simulate_data(
  n,
  J,
  b0 = NULL,
  b1 = NULL,
  distn,
  zinb_size = NULL,
  zinb_zero_prop = NULL,
  mean_z,
  X = NULL,
  B = NULL,
  cluster = NULL
)
}
\arguments{
\item{n}{Number of samples}

\item{J}{Number of categories}

\item{b0}{Intercept parameter vector}

\item{b1}{Covariate paramter vector}

\item{distn}{Distribution to simulate from, either "Poisson" or "ZINB"}

\item{zinb_size}{Size parameter for negative binomial draw for ZINB data}

\item{zinb_zero_prop}{Proportion of zeros for ZINB data}

\item{mean_z}{Parameter controlling the mean of the sample-specific effects.}

\item{X}{Optional design matrix, this must have two columns and n rows.}

\item{B}{Optional B matrix, if p is not equal to 2}

\item{cluster}{Optional cluster vector, this must have n elements.}
}
\value{
\code{Y}. A \code{n times J} dimension matrix of simulated response counts.
}
\description{
Function to simulate data for simulations in Clausen & Willis (2024) and for the cluster vignette
}
