% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_linear_combo.R
\name{test_linear_combo}
\alias{test_linear_combo}
\title{Test a linear combination}
\usage{
test_linear_combo(fitted_model, linear_combo, hypothesis = 0, j = NULL)
}
\arguments{
\item{fitted_model}{An \code{emuFit} fitted model.}

\item{linear_combo}{A vector of length \code{p = ncol(X)} to determine the linear
combination of parameters to be tested.}

\item{hypothesis}{The null hypothesis to compare the linear combination of
coefficients against. The default value is \code{0}.}

\item{j}{A vector of \code{j} values giving the indices of taxa to test. The default
behavior is to include all taxa.}
}
\value{
A data frame giving confidence intervals using robust standard errors, test
statistics for the robust Wald test, and p-values, for the linear combination of
parameters that is being tested, for each taxon specified in \code{j}.
}
\description{
Test a hypothesis about a linear combination of parameters. Currently this is
only implemented for robust Wald tests.
}
\examples{
data(wirbel_sample_small)
data(wirbel_otu_small)
emuRes <- emuFit(formula = ~ Group + Country, data = wirbel_sample_small, Y = wirbel_otu_small,
                 test_kj = data.frame(k = 2, j = 1), tolerance = 0.01, 
                 constraint_tol = 0.01, B_null_tol = 0.01) 
 # here we set large tolerances for the example to run quickly, 
 # but we recommend smaller tolerances in practice
# test whether CountryFRA - constraint_FRA = CountryUSA - constraint_USA for taxon 1 
linear_combo_res <- test_linear_combo(fitted_res = emuRes, 
                                     linear_combo = c(0, 0, 1, -1),
                                     j = 1)

}
