% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wirbel_otu.R
\docType{data}
\name{wirbel_otu}
\alias{wirbel_otu}
\title{mOTU table from Wirbel et al. meta-analysis of fecal metagenomes.}
\format{
A matrix with mOTU counts.
\describe{
\item{otu_table}{OTU table with 566 samples and 845 taxa}
}
}
\usage{
wirbel_otu
}
\description{
A matrix with counts of mOTUs (metagenome-derived OTUs) for each sample.
}
\references{
Wirbel, J et al. (2019). \emph{Meta-analysis of fecal metagenomes reveals global microbial signatures that are specific for colorectal cancer}. Nature Medicine, 25, 679–689. <doi: 10.1038/s41591-019-0406-6>.
}
\keyword{datasets}
