% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wirbel_taxonomy.R
\docType{data}
\name{wirbel_taxonomy}
\alias{wirbel_taxonomy}
\title{Taxonomy information for mOTUs from Wirbel et al. meta analysis.}
\format{
A matrix with taxonomy information.
\describe{
\item{tax_table}{taxonomy table}
}
}
\usage{
wirbel_taxonomy
}
\description{
A matrix with taxonomy information.
}
\references{
Wirbel, J et al. (2019). \emph{Meta-analysis of fecal metagenomes reveals global microbial signatures that are specific for colorectal cancer}. Nature Medicine, 25, 679–689. <doi: 10.1038/s41591-019-0406-6>.
}
\keyword{datasets}
