% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuFit.R
\name{emuFit}
\alias{emuFit}
\title{Fit radEmu model}
\usage{
emuFit(
  Y,
  formula,
  data,
  penalize = TRUE,
  B = NULL,
  test_kj = NULL,
  verbose = TRUE,
  tolerance = 0.01,
  maxit = 500,
  constraint_fn = pseudohuber_center,
  constraint_grad_fn = dpseudohuber_center_dx,
  constraint_param = 1,
  alpha = 0.05,
  return_wald_p = FALSE,
  run_score_tests = TRUE,
  rho_init = 1,
  rho_scaling = 2,
  gap_tolerance = 0.001
)
}
\arguments{
\item{Y}{an n x J matrix of nonnegative observations}

\item{formula}{a one-sided formula specifying the form of the mean model to be fit}

\item{data}{an n x p data frame containing variables given in \code{formula}}

\item{penalize}{logical: should Firth penalty be used in fitting model? Default is TRUE.}

\item{B}{starting value of coefficient matrix (p x J). If not provided,
B will be initiated as a zero matrix.}

\item{test_kj}{a data frame whose rows give coordinates (in category j and
covariate k) of elements of B to construct hypothesis tests for. If \code{test_kj}
is not provided, all elements of B save the intercept row will be tested.}

\item{verbose}{provide updates as model is being fitted? Defaults to TRUE.}

\item{tolerance}{tolerance on squared norm of gradient of likelihood to be used
as a stopping condition in model fitting}

\item{maxit}{maximum number of coordinate descent cycles to perform before
exiting optimization}

\item{constraint_fn}{function g defining a constraint on rows of B; g(B_k) = 0
for rows k = 1, ..., p of B. Default function is a smoothed median (minimizer of
pseudohuber loss).}

\item{constraint_grad_fn}{derivative of constraint_fn with respect to its
arguments (i.e., elements of a row of B)}

\item{constraint_param}{If pseudohuber centering is used (this is the default),
parameter controlling relative weighting of elements closer and further from center.
(Limit as \code{constraint_param} approaches infinity is the mean; as this parameter approaches zero,
the minimizer of the pseudo-Huber loss approaches the median.)}

\item{alpha}{nominal type 1 error level to be used to construct confidence intervals. Default is 0.05
(corresponding to 95\% confidence intervals)}

\item{return_wald_p}{logical: return p-values from Wald tests? Default is FALSE.}

\item{run_score_tests}{logical: perform robust score testing? Default is TRUE.}

\item{rho_init}{numeric: value at which to initiate rho parameter in augmented Lagrangian
algorithm. Default is 1.}

\item{rho_scaling}{numeric: value to scale rho by in each iteration of augmented Lagrangian
algorithm. Default is 2.}

\item{gap_tolerance}{numeric: constraint tolerance for fits under null hypotheses
(tested element of B must be equal to constraint function to within this tolerance for
a fit to be accepted as a solution to constrained optimization problem). Default is 1e-6.}
}
\value{
A dataframe containing point estimates, confidence intervals, and p-values for
fitted model
}
\description{
Fit radEmu model
}
