% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuFit.R
\name{emuFit}
\alias{emuFit}
\title{Fit radEmu model}
\usage{
emuFit(
  Y,
  formula,
  data,
  penalize = TRUE,
  B = NULL,
  test_kj = NULL,
  verbose = TRUE,
  tolerance = 0.01,
  maxit = 500,
  tol_factor = 0.01,
  min_tol = 0.01,
  constraint_fn = pseudohuber_center,
  constraint_grad_fn = dpseudohuber_center_dx,
  constraint_param = 1,
  alpha = 0.05,
  return_wald_p = FALSE,
  run_score_tests = TRUE,
  rho_init = 1,
  rho_scaling = 2,
  gap_tolerance = 1e-06
)
}
\arguments{
\item{Y}{an n x J matrix of nonnegative observations}

\item{formula}{a one-sided formula specifying the form of the mean model to be fit}

\item{data}{an n x p data frame containing variables given in \code{formula}}

\item{B}{starting value of coefficient matrix (p x J). If not provided,
B will be initiated as a zero matrix.}

\item{test_kj}{a data frame whose rows give coordinates (in category j and
covariate k) of elements of B to construct hypothesis tests for. If \code{test_kj}
is not provided, all elements of B save the intercept row will be tested.}

\item{verbose}{provide updates as model is being fitted? Defaults to TRUE.}

\item{tolerance}{tolerance on squared norm of gradient of likelihood to be used
as a stopping condition in model fitting}

\item{maxit}{maximum number of coordinate descent cycles to perform before
exiting optimization}

\item{constraint_fn}{function g defining a constraint on rows of B; g(B_k) = 0
for rows k = 1, ..., p of B. Default function is a smoothed median (minimizer of
pseudohuber loss).}

\item{constraint_grad_fn}{derivative of constraint_fn with respect to its
arguments (i.e., elements of a row of B)}

\item{constraint_param}{If pseudohuber centering is used (this is the default),
parameter controlling relative weighting of elements closer and further from center.
(Limit as \code{constraint_param} approaches infinity is the mean; as this parameter approaches zero,
the minimizer of the pseudo-Huber loss approaches the median.)}
}
\value{
A p x J matrix containing regression coefficients (under constraint
g(B_k) = 0)
}
\description{
Fit radEmu model
}
