% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuFit.R
\name{emuFit}
\alias{emuFit}
\title{Fit partially identified log-linear model to 'relative abundance' data}
\usage{
emuFit(
  formula_rhs = NULL,
  Y,
  X = NULL,
  covariate_data = NULL,
  B = NULL,
  B_cutoff = 20,
  tolerance = 0.1,
  maxit = 100,
  verbose = TRUE,
  constraint_fn = NULL,
  maxit_glm = 100,
  optim_only = TRUE,
  method = "ML",
  guarded = TRUE,
  linesearch = FALSE,
  reweight = FALSE,
  reweight_blocks = NULL,
  weights = NULL,
  test_firth = FALSE,
  return_a_lot = TRUE,
  prefit = TRUE
)
}
\arguments{
\item{formula_rhs}{The right-hand side of a formula specifying which
predictors should be included in model}

\item{Y}{A matrix of outcome data with rows corresponding to samples and
columns corresponding to outcome categories (e.g., microbial taxon)}

\item{constraint_fn}{The function to be used as a constraint -- namely,
we enforce constraint_fn(B_k) = 0 for all rows k = 1, ..., p of B. This
defaults to median().}

\item{reweight}{Refit model using weights derived from estimated mean-
variance relationship? Default is FALSE.}
}
\value{
\item{emuMod}{An emuMod object specifying the model fit and providing
point estimates for effects included in model.}
}
\description{
This function fits a radEmu model to multivariate outcome data Y. Predictors
are specified either via a formula, in which case covariate data must be
included as a data frame, or via an already formed model matrix X.
}
\author{
David Clausen
}
