% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuFit_check.R
\name{emuFit_check}
\alias{emuFit_check}
\title{Runs checks for appropriate arguments before running \code{emuFit()}}
\usage{
emuFit_check(
  Y,
  X = NULL,
  formula = NULL,
  data = NULL,
  assay_name = NULL,
  cluster = NULL,
  B_null_list = NULL,
  test_kj = NULL,
  match_row_names = TRUE,
  verbose = FALSE,
  remove_zero_comparison_pvals = 0.01,
  unobserved_taxon_error = TRUE
)
}
\arguments{
\item{Y}{an n x J matrix or dataframe of nonnegative observations, or a phyloseq object containing an otu table and sample data.}

\item{X}{an n x p matrix or dataframe of covariates (optional)}

\item{formula}{a one-sided formula specifying the form of the mean model to be fit}

\item{data}{an n x p data frame containing variables given in \code{formula}}

\item{assay_name}{a string containing the desired assay name within a \code{TreeSummarizedExperiment} object.
This is only required if Y is a \code{TreeSummarizedExperiment} object, otherwise this argument does nothing
and can be ignored.}

\item{cluster}{a vector giving cluster membership for each row of Y to be used in computing
GEE test statistics. Default is NULL, in which case rows of Y are treated as independent.}

\item{B_null_list}{list of starting values of coefficient matrix (p x J) for null estimation. This should either
be a list with the same length as \code{test_kj}. If you only want to provide starting values for some tests,
include the other elements of the list as \code{NULL}.}

\item{test_kj}{a data frame whose rows give coordinates (in category j and
covariate k) of elements of B to construct hypothesis tests for. If \code{test_kj}
is not provided, all elements of B save the intercept row will be tested.}

\item{match_row_names}{logical: Make sure rows on covariate data and response data correspond to
the same sample by comparing row names and subsetting/reordering if necessary.}

\item{verbose}{provide updates as model is being fitted? Defaults to FALSE. If user sets verbose = TRUE,
then key messages about algorithm progress will be displayed. If user sets verbose = "development",
then key messages and technical messages about convergence will be displayed. Most users who want status
updates should set verbose = TRUE.}

\item{remove_zero_comparison_pvals}{Should score p-values be replaced with NA for zero-comparison parameters? These parameters occur
for categorical covariates with three or more levels, and represent parameters that compare a covariate level to the reference level for
a category in which the comparison level and reference level both have 0 counts in all samples. These parameters can have misleadingly
small p-values and are not thought to have scientifically interesting signals. We recommend removing them before analyzing data further.
If TRUE, all zero-comparison parameter p-values will be set to NA. If FALSE no zero-comparison parameter p-values will be set to NA.
If a value between 0 and 1, all zero-comparison p-values below the value will be set to NA.
Default is \code{0.01}.}

\item{unobserved_taxon_error}{logical: should an error be thrown if Y includes taxa that have 0 counts for all samples? Default is TRUE.}
}
\value{
returns objects \code{Y}, \code{X}, \code{cluster}, and \code{B_null_list}, which may be modified by tests, and throw any useful
errors, warnings, or messages.
}
\description{
Runs checks for appropriate arguments before running \code{emuFit()}
}
