% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuFit_micro_penalized.R
\name{emuFit_micro_penalized}
\alias{emuFit_micro_penalized}
\title{Fit radEmu model with Firth penalty}
\usage{
emuFit_micro_penalized(
  X,
  Y,
  B = NULL,
  constraint_fn = NULL,
  maxit = 500,
  tolerance = 1e-05,
  max_step = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{X}{a p x J design matrix}

\item{Y}{an n x p matrix of nonnegative observations}

\item{B}{starting value of coefficient matrix (p x J)}

\item{constraint_fn}{function g defining constraint on rows of B; g(B_k) = 0
for rows k = 1, ..., p of B.}

\item{maxit}{maximum number of coordinate descent cycles to perform before
exiting optimization}

\item{tolerance}{tolerance on improvement in log likelihood at which to
exit optimization}
}
\value{
A p x J matrix containing regression coefficients (under constraint
g(B_k) = 0)
}
\description{
Fit radEmu model with Firth penalty
}
